Refactor: Consolidate shared adapter methods into dspxUtils, reduce redundant code
