+++ modules/dspxBidAdapter.js
@@ -1,9 +1,22 @@
-import {deepAccess, getBidIdParameter, isFn, logError, logMessage, logWarn, isEmptyStr, isArray} from '../src/utils.js';
-
+import {deepAccess, logMessage, getBidIdParameter, logError, logWarn} from '../src/utils.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import {BANNER, VIDEO} from '../src/mediaTypes.js';
-import {Renderer} from '../src/Renderer.js';
 import {includes} from '../src/polyfill.js';
+import {
+  fillUsersIds,
+  handleSyncUrls,
+  objectToQueryString,
+  isBannerRequest,
+  getVideoContext,
+  convertMediaInfoForRequest,
+  getMediaTypesInfo,
+  getBidFloor,
+  siteContentToString,
+  assignDefinedValues,
+  extractUserSegments,
+  interpretResponse
+} from '../libraries/dspxUtils/bidderUtils.js';
+import {Renderer} from '../src/Renderer.js';
 
 /**
  * @typedef {import('../src/adapters/bidderFactory.js').BidRequest} BidRequest
@@ -153,341 +166,11 @@ export const spec = {
   interpretResponse: function(serverResponse, bidRequest) {
     logMessage('DSPx: serverResponse', serverResponse);
     logMessage('DSPx: bidRequest', bidRequest);
-    const bidResponses = [];
-    const response = serverResponse.body;
-    const crid = response.crid || 0;
-    const cpm = response.cpm / 1000000 || 0;
-    if (cpm !== 0 && crid !== 0) {
-      const dealId = response.dealid || '';
-      const currency = response.currency || 'EUR';
-      const netRevenue = (response.netRevenue === undefined) ? true : response.netRevenue;
-      const bidResponse = {
-        requestId: response.bid_id,
-        cpm: cpm,
-        width: response.width,
-        height: response.height,
-        creativeId: crid,
-        dealId: dealId,
-        currency: currency,
-        netRevenue: netRevenue,
-        type: response.type,
-        ttl: 60,
-        meta: {
-          advertiserDomains: response.adomain || []
-        }
-      };
-
-      if (response.vastUrl) {
-        bidResponse.vastUrl = response.vastUrl;
-        bidResponse.mediaType = 'video';
-      }
-      if (response.vastXml) {
-        bidResponse.vastXml = response.vastXml;
-        bidResponse.mediaType = 'video';
-      }
-      if (response.renderer) {
-        bidResponse.renderer = newRenderer(bidRequest, response);
-      }
-
-      if (response.videoCacheKey) {
-        bidResponse.videoCacheKey = response.videoCacheKey;
-      }
-
-      if (response.adTag) {
-        bidResponse.ad = response.adTag;
-      }
-
-      if (response.bid_appendix) {
-        Object.keys(response.bid_appendix).forEach(fieldName => {
-          bidResponse[fieldName] = response.bid_appendix[fieldName];
-        });
-      }
-
-      bidResponses.push(bidResponse);
-    }
-    return bidResponses;
+    return interpretResponse(serverResponse, bidRequest, (bidRequest, response) => newRenderer(bidRequest, response));
   },
   getUserSyncs: function(syncOptions, serverResponses, gdprConsent, uspConsent) {
-    if (!serverResponses || serverResponses.length === 0) {
-      return [];
-    }
-
-    const syncs = []
-
-    let gdprParams = '';
-    if (gdprConsent) {
-      if ('gdprApplies' in gdprConsent && typeof gdprConsent.gdprApplies === 'boolean') {
-        gdprParams = `gdpr=${Number(gdprConsent.gdprApplies)}&gdpr_consent=${gdprConsent.consentString}`;
-      } else {
-        gdprParams = `gdpr_consent=${gdprConsent.consentString}`;
-      }
-    }
-
-    if (serverResponses.length > 0 && serverResponses[0].body.userSync) {
-      if (syncOptions.iframeEnabled) {
-        serverResponses[0].body.userSync.iframeUrl.forEach((url) => syncs.push({
-          type: 'iframe',
-          url: appendToUrl(url, gdprParams)
-        }));
-      }
-      if (syncOptions.pixelEnabled) {
-        serverResponses[0].body.userSync.imageUrl.forEach((url) => syncs.push({
-          type: 'image',
-          url: appendToUrl(url, gdprParams)
-        }));
-      }
-    }
-    return syncs;
+    return handleSyncUrls(syncOptions, serverResponses, gdprConsent);
   }
-}
-
-/**
- * Adds userIds to payload
- *
- * @param bidRequest
- * @param payload
- */
-function fillUsersIds(bidRequest, payload) {
-  if (bidRequest.hasOwnProperty('userId')) {
-    let didMapping = {
-      did_netid: 'userId.netId',
-      did_id5: 'userId.id5id.uid',
-      did_id5_linktype: 'userId.id5id.ext.linkType',
-      did_uid2: 'userId.uid2',
-      did_sharedid: 'userId.sharedid',
-      did_pubcid: 'userId.pubcid',
-      did_uqid: 'userId.utiq',
-      did_cruid: 'userId.criteoid',
-      did_euid: 'userId.euid',
-      // did_tdid: 'unifiedId',
-      did_tdid: 'userId.tdid',
-      did_ppuid: function() {
-        let path = 'userId.pubProvidedId';
-        let value = deepAccess(bidRequest, path);
-        if (isArray(value)) {
-          for (const rec of value) {
-            if (rec.uids && rec.uids.length > 0) {
-              for (let i = 0; i < rec.uids.length; i++) {
-                if ('id' in rec.uids[i] && deepAccess(rec.uids[i], 'ext.stype') === 'ppuid') {
-                  return (rec.uids[i].atype ?? '') + ':' + rec.source + ':' + rec.uids[i].id;
-                }
-              }
-            }
-          }
-        }
-        return undefined;
-      },
-      did_cpubcid: 'crumbs.pubcid'
-    };
-    for (let paramName in didMapping) {
-      let path = didMapping[paramName];
-
-      // handle function
-      if (typeof path == 'function') {
-        let value = path(paramName);
-        if (value) {
-          payload[paramName] = value;
-        }
-        continue;
-      }
-      // direct access
-      let value = deepAccess(bidRequest, path);
-      if (typeof value == 'string' || typeof value == 'number') {
-        payload[paramName] = value;
-      } else if (typeof value == 'object') {
-        // trying to find string ID value
-        if (typeof deepAccess(bidRequest, path + '.id') == 'string') {
-          payload[paramName] = deepAccess(bidRequest, path + '.id');
-        } else {
-          if (Object.keys(value).length > 0) {
-            logError(`DSPx: WARNING: fillUserIds had to use first key in user object to get value for bid.userId key: ${path}.`);
-            payload[paramName] = value[Object.keys(value)[0]];
-          }
-        }
-      }
-    }
-  }
-}
-
-function appendToUrl(url, what) {
-  if (!what) {
-    return url;
-  }
-  return url + (url.indexOf('?') !== -1 ? '&' : '?') + what;
-}
-
-function objectToQueryString(obj, prefix) {
-  let str = [];
-  let p;
-  for (p in obj) {
-    if (obj.hasOwnProperty(p)) {
-      let k = prefix ? prefix + '[' + p + ']' : p;
-      let v = obj[p];
-      str.push((v !== null && typeof v === 'object')
-        ? objectToQueryString(v, k)
-        : encodeURIComponent(k) + '=' + encodeURIComponent(v));
-    }
-  }
-  return str.filter(n => n).join('&');
-}
-
-/**
- * Check if it's a banner bid request
- *
- * @param {BidRequest} bid - Bid request generated from ad slots
- * @returns {boolean} True if it's a banner bid
- */
-function isBannerRequest(bid) {
-  return bid.mediaType === 'banner' || !!deepAccess(bid, 'mediaTypes.banner') || !isVideoRequest(bid);
-}
-
-/**
- * Check if it's a video bid request
- *
- * @param {BidRequest} bid - Bid request generated from ad slots
- * @returns {boolean} True if it's a video bid
- */
-function isVideoRequest(bid) {
-  return bid.mediaType === 'video' || !!deepAccess(bid, 'mediaTypes.video');
-}
-
-/**
- * Get video sizes
- *
- * @param {BidRequest} bid - Bid request generated from ad slots
- * @returns {object}
- */
-function getVideoSizes(bid) {
-  return parseSizes(deepAccess(bid, 'mediaTypes.video.playerSize') || bid.sizes);
-}
-
-/**
- * Get video context
- *
- * @param {BidRequest} bid - Bid request generated from ad slots
- * @returns {object}
- */
-function getVideoContext(bid) {
-  return deepAccess(bid, 'mediaTypes.video.context') || 'unknown';
-}
-
-/**
- * Get banner sizes
- *
- * @param {BidRequest} bid - Bid request generated from ad slots
- * @returns {object} True if it's a video bid
- */
-function getBannerSizes(bid) {
-  return parseSizes(deepAccess(bid, 'mediaTypes.banner.sizes') || bid.sizes);
-}
-
-/**
- * Parse size
- * @param size
- * @returns {object} sizeObj
- */
-function parseSize(size) {
-  let sizeObj = {}
-  sizeObj.width = parseInt(size[0], 10);
-  sizeObj.height = parseInt(size[1], 10);
-  return sizeObj;
-}
-
-/**
- * Parse sizes
- * @param sizes
- * @returns {{width: number , height: number }[]}
- */
-function parseSizes(sizes) {
-  if (Array.isArray(sizes[0])) { // is there several sizes ? (ie. [[728,90],[200,300]])
-    return sizes.map(size => parseSize(size));
-  }
-  return [parseSize(sizes)]; // or a single one ? (ie. [728,90])
-}
-
-/**
- * Get MediaInfo object for server request
- *
- * @param mediaTypesInfo
- * @returns {*}
- */
-function convertMediaInfoForRequest(mediaTypesInfo) {
-  let requestData = {};
-  Object.keys(mediaTypesInfo).forEach(mediaType => {
-    requestData[mediaType] = mediaTypesInfo[mediaType].map(size => {
-      return size.width + 'x' + size.height;
-    }).join(',');
-  });
-  return requestData;
-}
-
-/**
- * Get media types info
- *
- * @param bid
- */
-function getMediaTypesInfo(bid) {
-  let mediaTypesInfo = {};
-
-  if (bid.mediaTypes) {
-    Object.keys(bid.mediaTypes).forEach(mediaType => {
-      if (mediaType === BANNER) {
-        mediaTypesInfo[mediaType] = getBannerSizes(bid);
-      }
-      if (mediaType === VIDEO) {
-        mediaTypesInfo[mediaType] = getVideoSizes(bid);
-      }
-    });
-  } else {
-    mediaTypesInfo[BANNER] = getBannerSizes(bid);
-  }
-  return mediaTypesInfo;
-}
-
-/**
- * Get Bid Floor
- * @param bid
- * @returns {number|*}
- */
-function getBidFloor(bid) {
-  if (!isFn(bid.getFloor)) {
-    return deepAccess(bid, 'params.bidfloor', 0);
-  }
-
-  try {
-    const bidFloor = bid.getFloor({
-      currency: 'EUR',
-      mediaType: '*',
-      size: '*',
-    });
-    return bidFloor.floor;
-  } catch (_) {
-    return 0
-  }
-}
-
-/**
- * Create a new renderer
- *
- * @param bidRequest
- * @param response
- * @returns {Renderer}
- */
-function newRenderer(bidRequest, response) {
-  logMessage('DSPx: newRenderer', bidRequest, response);
-  const renderer = Renderer.install({
-    id: response.renderer.id || response.bid_id,
-    url: (bidRequest.params && bidRequest.params.rendererUrl) || response.renderer.url,
-    config: response.renderer.options || deepAccess(bidRequest, 'renderer.options'),
-    loaded: false
-  });
-
-  try {
-    renderer.setRender(outstreamRender);
-  } catch (err) {
-    logWarn('Prebid Error calling setRender on renderer', err);
-  }
-  return renderer;
 }
 
 /**
@@ -496,7 +179,7 @@ function newRenderer(bidRequest, response) {
  * @param bid
  */
 function outstreamRender(bid) {
-  logMessage('DSPx: outstreamRender bid:', bid);
+  logMessage('[DSPx][outstreamRender] bid:', bid);
   const embedCode = createOutstreamEmbedCode(bid);
   try {
     const inIframe = getBidIdParameter('iframe', bid.renderer.config);
@@ -507,7 +190,7 @@ function outstreamRender(bid) {
       if (typeof window.dspxRender === 'function') {
         window.dspxRender(bid);
       } else {
-        logError('[dspx][renderer] Error: dspxRender function is not found');
+        logError('[DSPx][outstreamRender] Error: dspxRender function is not found');
       }
       return;
     }
@@ -518,13 +201,13 @@ function outstreamRender(bid) {
       if (typeof window.dspxRender === 'function') {
         window.dspxRender(bid);
       } else {
-        logError('[dspx][renderer] Error: dspxRender function is not found');
+        logError('[DSPx][outstreamRender] Error: dspxRender function is not found');
       }
     } else if (slot) {
-      logError('[dspx][renderer] Error: slot not found');
+      logError('[DSPx][outstreamRender] Error: slot not found');
     }
   } catch (err) {
-    logError('[dspx][renderer] Error:' + err.message)
+    logError('[DSPx][outstreamRender] Error:' + err.message)
   }
 }
 
@@ -561,73 +244,27 @@ function createOutstreamEmbedCode(bid) {
 }
 
 /**
- * Convert site.content to string
- * @param content
+ * Create a new renderer
+ *
+ * @param bidRequest
+ * @param response
+ * @returns {Renderer}
  */
-function siteContentToString(content) {
-  if (!content) {
-    return '';
-  }
-  let stringKeys = ['id', 'title', 'series', 'season', 'artist', 'genre', 'isrc', 'url', 'keywords'];
-  let intKeys = ['episode', 'context', 'livestream'];
-  let arrKeys = ['cat'];
-  let retArr = [];
-  arrKeys.forEach(k => {
-    let val = deepAccess(content, k);
-    if (val && Array.isArray(val)) {
-      retArr.push(k + ':' + val.join('|'));
-    }
+function newRenderer(bidRequest, response) {
+  logMessage('[DSPx] newRenderer', bidRequest, response);
+  const renderer = Renderer.install({
+    id: response.renderer.id || response.bid_id,
+    url: (bidRequest.params && bidRequest.params.rendererUrl) || response.renderer.url,
+    config: response.renderer.options || deepAccess(bidRequest, 'renderer.options'),
+    loaded: false
   });
-  intKeys.forEach(k => {
-    let val = deepAccess(content, k);
-    if (val && typeof val === 'number') {
-      retArr.push(k + ':' + val);
-    }
-  });
-  stringKeys.forEach(k => {
-    let val = deepAccess(content, k);
-    if (val && typeof val === 'string') {
-      retArr.push(k + ':' + encodeURIComponent(val));
-    }
-  });
-  return retArr.join(',');
-}
 
-/**
- * Assigns multiple values to the specified keys on an object if the values are not undefined.
- * @param {Object} target - The object to which the values will be assigned.
- * @param {Object} values - An object containing key-value pairs to be assigned.
- */
-function assignDefinedValues(target, values) {
-  for (const key in values) {
-    if (values[key] !== undefined) {
-      target[key] = values[key];
-    }
+  try {
+    renderer.setRender(outstreamRender);
+  } catch (err) {
+    logWarn('[DSPx]Prebid Error calling setRender on renderer', err);
   }
-}
-
-/**
- * Extracts user segments/topics from the bid request object
- * @param {Object} bid - The bid request object
- * @returns {{segclass: *, segtax: *, segments: *}|undefined} - User segments/topics or undefined if not found
- */
-function extractUserSegments(bid) {
-  const userData = deepAccess(bid, 'ortb2.user.data') || [];
-  for (const dataObj of userData) {
-    if (dataObj.segment && isArray(dataObj.segment) && dataObj.segment.length > 0) {
-      const segments = dataObj.segment
-        .filter(seg => seg.id && !isEmptyStr(seg.id) && isFinite(seg.id))
-        .map(seg => Number(seg.id));
-      if (segments.length > 0) {
-        return {
-          segtax: deepAccess(dataObj, 'ext.segtax'),
-          segclass: deepAccess(dataObj, 'ext.segclass'),
-          segments: segments.join(',')
-        };
-      }
-    }
-  }
-  return undefined;
+  return renderer;
 }
 
 registerBidder(spec);
